Create a package to house types used at build time but not at runtime and relocate model parser types there
