+++ packages/build-types/.gitignore
@@ -1,0 +1,7 @@
+/node_modules/
+/build/
+/coverage/
+/docs/
+*.tgz
+*.log
+package-lock.json
