+++ packages/build-types/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "@aws/build-types",
+    "version": "0.0.1",
+    "scripts": {
+        "prepublishOnly": "tsc",
+        "pretest": "tsc",
+        "test": "exit 0"
+    },
+    "main": "./build/index.js",
+    "types": "./build/index.d.ts",
+    "author": "aws-sdk-js@amazon.com",
+    "license": "Apache-2.0",
+    "dependencies": {
+        "@aws/types": "^0.0.1"
+    },
+    "devDependencies": {
+        "typescript": "^2.3"
+    }
+}
