+++ packages/service-model/package.json
@@ -12,6 +12,7 @@
     "author": "aws-javascript-sdk-team@amazon.com",
     "license": "UNLICENSED",
     "dependencies": {
+        "@aws/build-types": "^0.0.1",
         "@aws/types": "^0.0.1"
     },
     "devDependencies": {
