+++ packages/service-types-generator/src/Components/Type/Output.spec.ts
@@ -5,7 +5,7 @@ import {
     TreeModelList,
     TreeModelMap,
     TreeModelStructure,
-} from "@aws/service-model";
+} from "@aws/build-types";
 import {getMemberType} from "./getMemberType";
 import {
     NonStreamingBlob,
