+++ packages/service-types-generator/src/Components/Type/Structure.spec.ts
@@ -6,7 +6,7 @@ import {ModeledStructure} from "./ModeledStructure";
 import {
     TreeModelShape,
     TreeModelStructure,
-} from "@aws/service-model";
+} from "@aws/build-types";
 
 describe('Structure', () => {
     it('should mark required members as not optional', () => {
