+++ packages/service-types-generator/src/Components/Type/getMapIndexDeclaration.ts
@@ -1,5 +1,5 @@
 import {getStringDeclaration} from "./getStringDeclaration";
-import {TreeModelMap, TreeModelShape} from "@aws/service-model";
+import {TreeModelMap} from "@aws/build-types";
 
 export function getMapIndexDeclaration(map: TreeModelMap): string {
     if (map.key.shape.type !== 'string') {
