+++ packages/xml-builder/package.json
@@ -8,10 +8,6 @@
         "jest": "^20.0.4",
         "typescript": "^2.3"
     },
-    "dependencies": {
-        "@aws/service-model": "^0.0.1",
-        "@aws/types": "^0.0.1"
-    },
     "scripts": {
         "prepublishOnly": "tsc",
         "pretest": "tsc -p tsconfig.test.json",
